Updated pubspec dependencies. Extracted widgets. Prevent adding the same images.
