+++ pubspec.yaml
@@ -6,7 +6,7 @@ dependencies:
     sdk: flutter
   cupertino_icons: ^0.1.2
   groovin_material_icons: ^1.0.2
-  groovin_widgets: ^1.1.0
+  groovin_widgets: ^1.3.2
   rounded_modal: ^1.0.0
   outline_material_icons: ^0.1.0
   dynamic_theme: ^1.0.0
@@ -21,6 +21,7 @@ dependencies:
   flutter_slidable: "^0.4.9"
   rxdart: ^0.19.0
   permission_handler: ^1.0.1
+  photo_view: ^0.1.3
 
 dev_dependencies:
   flutter_test:
